Remove support for dynamically downloading distribution ZIP (#1184)
When building an image, the distribution ZIP can come from one of
several places:
1. from a file on the local filesystem
2. downloaded from a supplied URL
3. downloaded dynamically based on the latest released version

As discussed in https://github.com/hazelcast/hazelcast-docker/pull/1183,
only (1) is now actually used in our CI pipeline or for local
development, (2) & (3) aren't tested and an error has crept in.

Rather than fix the error, an _alternative_ solution is to remove the
(now unused) functionality. The _final image_ remains the same, but the
codebase is simpler.

Closes https://github.com/hazelcast/hazelcast-docker/pull/1183